Rename test suite to i2c_tests, add simple i2c_master_slave example
Move the hardware test suite to i2c_tests.rs and create a clean
i2c_master_slave.rs example showing basic master write, slave receive,
slave transmit, and write_read with repeated START. The example targets
the v2 slave driver only, with no internal cfg gates.
